+++ assets/js/panelbar.js
@@ -3,13 +3,13 @@
 
 var hasClass = function (elem, className) {
   return new RegExp(' ' + className + ' ').test(' ' + elem.className + ' ');
-}
+};
 
 var addClass = function (elem, className) {
   if (!hasClass(elem, className)) {
     elem.className += ' ' + className;
   }
-}
+};
 
 var removeClass = function (elem, className) {
   var newClass = ' ' + elem.className.replace( /[\t\r\n]/g, ' ') + ' ';
@@ -19,11 +19,12 @@ var removeClass = function (elem, className) {
     }
     elem.className = newClass.replace(/^\s+|\s+$/g, '');
   }
-}
+};
 
 
 // Elements
-var panelbar  = document.getElementById('panelbar');
+var wrapper   = document.getElementById('panelbar');
+var panelbar  = document.getElementById('panelbar_bar');
 var controls  = document.getElementById('panelbar_controls');
 var switchbtn = document.getElementById('panelbar_switch');
 var flipbtn   = document.getElementById('panelbar_flip');
@@ -33,25 +34,25 @@ var flipbtn   = document.getElementById('panelbar_flip');
 if ( 'querySelector' in document && 'addEventListener' in window ) {
 
   // Visibility toggle & flip
-  switchbtn.addEventListener('click', function (e) {
-    if (hasClass(panelbar, 'hidden')) {
-      removeClass(panelbar, 'hidden');
+  switchbtn.addEventListener('click', function () {
+    if (hasClass(panelbar, 'panelbar__bar--hidden')) {
+      removeClass(panelbar, 'panelbar__bar--hidden');
     } else {
-      addClass(panelbar, 'hidden');
+      addClass(panelbar, 'panelbar__bar--hidden');
     }
   });
 
-  flipbtn.addEventListener('click', function (e) {
-    if (hasClass(panelbar, 'top')) {
-      removeClass(panelbar, 'top');
+  flipbtn.addEventListener('click', function () {
+    if (hasClass(wrapper, 'panelbar--top')) {
+      removeClass(wrapper, 'panelbar--top');
     } else {
-      addClass(panelbar, 'top');
+      addClass(wrapper, 'panelbar--top');
     }
 
-    if (hasClass(panelbar, 'bottom')) {
-      removeClass(panelbar, 'bottom');
+    if (hasClass(wrapper, 'panelbar--bottom')) {
+      removeClass(wrapper, 'panelbar--bottom');
     } else {
-      addClass(panelbar, 'bottom');
+      addClass(wrapper, 'panelbar--bottom');
     }
 
   });
@@ -61,37 +62,5 @@ if ( 'querySelector' in document && 'addEventListener' in window ) {
   // remove switch in legacy Browser
   controls.remove();
   panelbar.style.paddingRight = 0;
-  panelbar.classList.remove("hidden");
-}
-
-
-
-// EnhancedJS with jQuery
-
-if (jQuery && enhancedJS === true) {
-  $(function() {
-
-    // Element: toggle
-    $(".panelbar--toggle > a").on("click", function (e) {
-      e.preventDefault();
-
-      var status = $(this).find('span').text() == 'Visible' ? 'hide' : 'publish';
-      var url    = siteURL + "/panel/api/pages/" + status + "/" + currentURI;
-
-      $.ajax({
-        type: "POST",
-        url: url
-      });
-
-      $(this).find('.fa').toggleClass('fa-toggle-off fa-toggle-on');
-      $(this).find('span').text(status == "hide" ? "Invisible" : "Visible");
-
-      setTimeout(function() {
-        location.reload();
-      }, 100);
-
-    });
-
-
-  });
+  panelbar.classList.remove("panelbar--hidden");
 }